# src/main/resources/static/swagger-api/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "ecommerce-spring-reactjs"
servers:
  - url: "http://localhost:8080"
  - url: "https://perfume-websore-api.herokuapp.com"
...
paths:
...
  /api/v1/perfumes/ids:
    post:
      tags:
        - "PerfumeController"
      summary: "Get perfumes by ids"
      operationId: "getPerfumesByIds"
      requestBody:
        content:
          'application/json':
            schema:
              type: "array"
              items:
                type: "integer"
                format: int64
              example: [1, 2, 3]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by filter parameters"
      operationId: "findPerfumesByFilterParams"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                perfumers:
                  type: "array"
                  items:
                    type: "string"
                    example: "Dior"
                genders:
                  type: "array"
                  items:
                    type: "string"
                    example: "male"
                prices:
                  type: "array"
                  items:
                    type: "integer"
                    format: int32
                  example: [1, 200]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search/gender:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by gender"
      operationId: "findByPerfumeGender"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                genders:
                  type: "array"
                  items:
                    type: "string"
                    example: "male"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search/perfumer:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by brand"
      operationId: "findByPerfumer"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                perfumers:
                  type: "array"
                  items:
                    type: "string"
                  example: ["Dior", "Creed"]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/graphql/ids:
    post:
      tags:
        - "PerfumeController"
      summary: "Get list of perfumes by ids GraphQL query"
      operationId: "getPerfumesByIdsQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfumesIds(ids: [16,17,18]) {id perfumeTitle perfumer price filename} }"

      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    perfumeTitle:
                      type: "string"
                    perfumer:
                      type: "string"
                    price:
                      type: "integer"
                      format: int32
                    filename:
                      type: "string"
                  example:
                    id: 43
                    perfumeTitle: "Sauvage"
                    perfumer: "Dior"
                    price: 62
                    filename: "empty.jpg"
  /api/v1/perfumes/graphql/perfumes:
    post:
      tags:
        - "PerfumeController"
      summary: "Get list of perfumes by GraphQL query"
      operationId: "getAllPerfumesByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfumes {id perfumeTitle perfumer price filename} }"

      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    perfumeTitle:
                      type: "string"
                    perfumer:
                      type: "string"
                    price:
                      type: "integer"
                      format: int32
                    filename:
                      type: "string"
                  example:
                    id: 43
                    perfumeTitle: "Sauvage"
                    perfumer: "Dior"
                    price: 62
                    filename: "empty.jpg"
  /api/v1/perfumes/graphql/perfume:
    post:
      tags:
        - "PerfumeController"
      summary: "Get perfume by GraphQL query"
      operationId: "getPerfumeByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfume(id: 43) {id perfumeTitle perfumer price filename} }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    format: int64
                  perfumeTitle:
                    type: "string"
                  perfumer:
                    type: "string"
                  price:
                    type: "integer"
                    format: int32
                  filename:
                    type: "string"
                example:
                  id: 43
                  perfumeTitle: "Sauvage"
                  perfumer: "Dior"
                  price: 62
                  filename: "empty.jpg"
  /api/v1/registration:
    post:
      tags:
        - "RegistrationController"
      summary: "Registration"
      operationId: "registration"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: "string"
                  example: "test1234@test.com"
                password:
                  type: "string"
                  example: "1234567890"
                password2:
                  type: "string"
                  example: "1234567890"
                firstName:
                  type: "string"
                  example: "John"
                lastName:
                  type: "string"
                  example: "Doe"
                captcha:
                  type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: "string"
                    example: "User successfully registered."
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                PasswordMatchError:
                  summary: "Password Match Error"
                  value:
                    passwordError: "Passwords do not match."
                EmailError:
                  summary: "Email Error"
                  value:
                    emailError: "Email is already used."
                BindingResultError:
                  summary: "Binding Result Errors"
                  value:
                    captchaError: "Fill captcha."
                    firstNameError: "First name cannot be empty"
                    lastNameError: "Last name cannot be empty"
                    passwordError: "The password must be between 6 and 16 characters long"
                    password2Error: "The password confirmation must be between 6 and 16 characters long"
                    emailError: "Incorrect email"
  /api/v1/registration/activate/{code}:
    get:
      tags:
        - "RegistrationController"
      summary: "Activate email code"
      operationId: "activateEmailCode"
      parameters:
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: "string"
                    example: "User successfully activated."
        "404":
          description: "Not Found"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Activation code not found."
...
  /api/v1/users/order/{orderId}:
    get:
      tags:
        - "UserController"
      summary: "Get Order By Id"
      operationId: "getOrderById"
      parameters:
        - name: "orderId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: "Not Found"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Order not found."
      security:
        - token: []
  /api/v1/users/order/{orderId}/items:
    get:
      tags:
        - "UserController"
      summary: "Get Order Items By Order Id"
      operationId: "getOrderItemsByOrderId"
      parameters:
        - name: "orderId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderItemResponse"
        "404":
          description: "Not Found"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Order not found."
      security:
        - token: [ ]
...
  /api/v1/user/graphql/orders:
    post:
      tags:
        - "UserController"
      summary: "Get all user orders by GraphQL query"
      operationId: "getUserOrdersByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ ordersByEmail(email: test123@test.com) { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/users/order:
    post:
      tags:
        - "UserController"
      summary: "Post order"
      operationId: "postOrder"
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                totalPrice:
                  type: "number"
                  format: "double"
                  example: 100
                firstName:
                  type: "string"
                  example: "John"
                lastName:
                  type: "string"
                  example: "Doe"
                city:
                  type: "string"
                  example: "New York"
                address:
                  type: "string"
                  example: "Wall Street 1"
                email:
                  type: "string"
                  example: "test123@test.com"
                phoneNumber:
                  type: "string"
                  example: "123456"
                postIndex:
                  type: "integer"
                  format: "int32"
                  example: 123456
                perfumesId:
                  type: "object"
                  properties:
                    43:
                      type: integer
                      example: 1
              required:
                - firstName
                - lastName
                - city
                - address
                - email
                - phoneNumber
                - postIndex
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  firstNameError:
                    type: "string"
                    example: "Fill in the input field"
                  lastNameError:
                    type: "string"
                    example: "Fill in the input field"
                  cityError:
                    type: "string"
                    example: "Fill in the input field"
                  addressError:
                    type: "string"
                    example: "Fill in the input field"
                  emailError:
                    type: "string"
                    example: "Email cannot be empty"
                  phoneNumberError:
                    type: "string"
                    example: "Phone number cannot be empty"
                  postIndexError:
                    type: "string"
                    example: "Post index cannot be empty"
...
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    PerfumeResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 43
        perfumeTitle:
          type: "string"
          example: "Sauvage"
        perfumer:
          type: "string"
          example: "Dior"
        price:
          type: "integer"
          format: "int32"
          example: 62
        filename:
          type: "string"
          example: "Dior_Sauvage.jpg"
        perfumeRating:
          type: "integer"
          format: "int32"
          example: 5
...
    OrderResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        totalPrice:
          type: "number"
          format: "double"
          example: 100
        date:
          type: "string"
          example: "2021-04-07"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        city:
          type: "string"
          example: "New York"
        address:
          type: "string"
          example: "Wall Street 1"
        email:
          type: "string"
          example: "test123@test.com"
        phoneNumber:
          type: "string"
          example: "123456"
        postIndex:
          type: "integer"
          format: "int32"
          example: 123456
    OrderItemResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        amount:
          type: "integer"
          format: "int64"
          example: 100
        quantity:
          type: "integer"
          format: "int64"
          example: 2
        perfume:
          $ref: "#/components/schemas/PerfumeResponse"
...